requirements.txt
opencv-python
torch
ultralytics
matplotlib
numpy
Pillow
torchvision
pyserial
cvzone
